fixed mutation: values exceeding range part2

+++ evaluation_functions.py
@@ -6,9 +6,9 @@ from individual import Individual
 def rosenbrock(individual):  # minimum szukamy
     result = numpy.float64(0)
     for i in range(len(individual) - 1):  # len(individual) = dimensions
-        result += 100 * (individual[i + 1] - individual[i] ** 2) ** 2 + (individual[i] - 1) ** 2
+        result += 100.0 * (individual[i + 1] - individual[i] ** 2) ** 2 + (individual[i] - 1) ** 2
 
-    return numpy.float64(result)
+    return result
 
 
 def zakharov(individual):  # minimum szukamy
@@ -18,14 +18,14 @@ def zakharov(individual):  # minimum szukamy
         result_1 += individual[i] ** 2
         result_2 += 0.5 * (i + 1) * individual[i]  # (i+1) - because we are starting from i = 0 in programming
 
-    return numpy.float64(result_1 + (result_2 ** 2) + (result_2 ** 4))
+    return result_1 + (result_2 ** 2) + (result_2 ** 4)
 
 
 def rastrigin(individual):
     temp = numpy.float64(0)
     for i in range(len(individual)):  # len(individual) = dimensions
         temp += individual[i] ** 2 - 10 * math.cos(2 * math.pi * individual[i])
-    return numpy.float64(10 * len(individual) + temp)  # len(individual) = dimensions
+    return 10 * len(individual) + temp  # len(individual) = dimensions
 
 
 def styblinski_tang(individual):
@@ -33,4 +33,4 @@ def styblinski_tang(individual):
     for i in range(len(individual)):
         temp += individual[i] ** 4 - 16 * individual[i] ** 2 + 5 * individual[i]
 
-    return numpy.float64(0.5 * temp)
+    return 0.5 * temp
